moved to visual studio 2015, fixed dependency system, TG-143 TG-144 #closed

+++ source/game/panels/LoadScreen.cpp
@@ -18,12 +18,12 @@ void LoadScreen::Draw(ControlDrawData*)
 	GUI.DrawSprite(tLoadbarBg, pt);
 
 	// loadbar
-	RECT r = {pt.x, pt.y, pt.x+8+int(progress*(503-8)), pt.y+desc.Height};
-	RECT rp = {0, 0, 8+int(progress*(503-8)), desc.Height};
+	RECT r = {pt.x, pt.y, pt.x+8+int(progress*(503-8)), LONG(pt.y+desc.Height)};
+	RECT rp = {0, 0, 8+int(progress*(503-8)), LONG(desc.Height)};
 	GUI.DrawSpriteRectPart(tLoadbar, r, rp);
 
 	// text
-	RECT r2 = {32, 0, GUI.wnd_size.x-32, GUI.wnd_size.y-desc.Height-32};
+	RECT r2 = {32, 0, GUI.wnd_size.x-32, LONG(GUI.wnd_size.y-desc.Height-32)};
 	GUI.DrawText(GUI.default_font, text, DT_CENTER|DT_BOTTOM, WHITE, r2);
 }
